Update CloudJump2.py to add Sprite class
Sprites will greatly simplify updating and drawing code by moving to scene.Layers.  I have started with just GrassBlocks in this pull request but in the future we can also migrate Enemy, Player, and most interestingly Cloud to be subclasses of Sprite.  There are some curious issues with Clouds that I believe will be cured by migrating them to be Sprites.

Interestingly, Layers have frames while Scenes have bounds so I changed most but not all instances of bounds to frame.

Rewrote `create_ground()` to take the desired number of grass blocks and return the ground level.  Modified the player creation logic to be at ground level on x and dead center on y.