Create cloud_maker.py to experiment w/ PIL clouds
The cloud images created by this script are images which simplifies coding of the game but they are `really ugly` compared to the ones created in CloudJump2.py.  I am lousy at PIL graphics coding so it would be a huge help if someone could make these clouds look at least as beautiful as the originals.