Create etude_on_player_death.py
Etude on Player Death - http://en.m.wikipedia.org/wiki/etude
Request for help on issue #9

The player needs to die with more panache to give our game an
arcade feel.  we could use your help on bot the sounds and the
animations around the Player.die() method.  Better noises and 
a player the shrinks to almost nothing and then is replace
with a puff of somke would be super cool. Thanks for playing!