Load images once at startup, rewrote draw() method
I have to say that I like it more when `wimpy_smoke` is set to False ;-).  I always find that it makes me laugh.  How about this as a compromise... When it is a normal end of game then wimpy_smoke=True but if the player sets a new high score then wimpy_smoke=False?

Future enhancement: For performance reasons and to allow people can play while not connected to the Internet, we should download and store images and sounds locally on the initial run.